+++ test/smokes/flake8/expectations.rb
@@ -194,7 +194,8 @@ Smoke.add_test(
     guid: "test-guid",
     timestamp: :_,
     type: "failure",
-    message: "The value of the attribute `$.linter.flake8.plugins` of `sideci.yml` is invalid.",
+    message:
+      "The value of the attribute `$.linter.flake8.plugins` in your `sideci.yml` is invalid. Please fix and retry.",
     analyzer: nil
   }
 )