+++ test/smokes/goodcheck/expectations.rb
@@ -76,7 +76,8 @@ Smoke.add_test(
     guid: "test-guid",
     timestamp: :_,
     type: "failure",
-    message: "The value of the attribute `$.linter.goodcheck.config` of `sideci.yml` is invalid.",
+    message:
+      "The value of the attribute `$.linter.goodcheck.config` in your `sideci.yml` is invalid. Please fix and retry.",
     analyzer: nil
   }
 )