+++ test/smokes/javasee/expectations.rb
@@ -60,7 +60,8 @@ Smoke.add_test(
     timestamp: :_,
     type: "failure",
     analyzer: nil,
-    message: "The value of the attribute `$.linter.javasee.dir[2]` of `sider.yml` is invalid."
+    message:
+      "The value of the attribute `$.linter.javasee.dir[2]` in your `sider.yml` is invalid. Please fix and retry."
   }
 )