+++ test/smokes/jshint/expectations.rb
@@ -82,7 +82,8 @@ Smoke.add_test(
     timestamp: :_,
     type: "failure",
     analyzer: nil,
-    message: "The value of the attribute `$.linter.jshint.config` of `sideci.yml` is invalid."
+    message:
+      "The value of the attribute `$.linter.jshint.config` in your `sideci.yml` is invalid. Please fix and retry."
   }
 )