+++ test/smokes/phpmd/expectations.rb
@@ -127,7 +127,8 @@ Smoke.add_test(
     guid: "test-guid",
     timestamp: :_,
     type: "failure",
-    message: "The value of the attribute `$.linter.phpmd.minimumpriority` of `sideci.yml` is invalid.",
+    message:
+      "The value of the attribute `$.linter.phpmd.minimumpriority` in your `sideci.yml` is invalid. Please fix and retry.",
     analyzer: nil
   }
 )