+++ test/smokes/pmd_java/expectations.rb
@@ -256,7 +256,8 @@ Smoke.add_test(
     timestamp: :_,
     type: "failure",
     analyzer: nil,
-    message: "The value of the attribute `$.linter.pmd_java.min_priority` of `sideci.yml` is invalid."
+    message:
+      "The value of the attribute `$.linter.pmd_java.min_priority` in your `sideci.yml` is invalid. Please fix and retry."
   }
 )