+++ test/smokes/stylelint/expectations.rb
@@ -527,7 +527,8 @@ Smoke.add_test(
     guid: "test-guid",
     timestamp: :_,
     type: "failure",
-    message: "The value of the attribute `$.linter.stylelint.options.ignore-path` of `sideci.yml` is invalid."
+    message:
+      "The value of the attribute `$.linter.stylelint.options.ignore-path` in your `sideci.yml` is invalid. Please fix and retry."
   }
 )